requirements.txt
flask
flask-cors
requests
python-dotenv
transformers
torch